// src/lib/internal/models/UserPrivateAccessTokenWithoutToken.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { UserAgent } from './UserAgent';

export type UserPrivateAccessTokenWithoutToken = {
    userid: number;
    refresh_token: string;
    expired_at: number;
    created_at: number;
    active: boolean;
    appid?: number;
    scope: string;
    updated_at: number;
    user_agent?: UserAgent;
};